Allow to specify a test run, without uploading/building any data, using existing bits
using the scenario of a localFolder upload has the consequence of
when running a test over and over again, the components have to be uploaded all the time.
this can be many MBs of data, slowing down the startup time.

with this change, we allow the job to specify that no data has to be built/uploaded, as it already exists.
one can write a job that uploades some binaries and later on, multiple jobs use these binaries.

an example job would be:

  UploadUtil:
    source:
      localFolder: c:\folderToUpload
    executable: c:\windows\system32\cmd.exe
    arguments: "/c xcopy /s /y /q . c:\\myUtil"

and later on using this:

  MyUtil:
    source:
      ExistingAppFolder: c:\\myUtil
    executable: c:\windows\system32\cmd.exe
    arguments: "/c myutil.cmd"